+++ src/extra/utils/misc.py
@@ -204,3 +204,105 @@ def hyperslicer2(arr, *args, ax=None, lognorm=False, colorbar=True, **kwargs):
         fig.colorbar(ax.get_images()[-1], ax=ax)
 
     return controls
+
+
+def ridgeplot(
+        data, *, fig=None, overlap=0.5, xlabel=None, ylabel="Per-line values",
+        ylim=None, yline=None, stack_label=None, stack_ticklabels=None
+):
+    """Make a ridgeline plot showing a sequence of similar lines
+
+    A ridgeline plot spreads out the different lines vertically to make their
+    order clear, but allowing them to overlap. It's an alternative to a heatmap,
+    especially if there are relatively few rows (around 5-20).
+
+    Args:
+        data (array_like): A 2D array, each row of which will be plotted as one
+            line, starting at the top of the plot. Pass an xarray DataArray to
+            use its labels by default.
+        fig (matplotlib.figure.Figure): Plot into an existing matplotlib figure.
+        overlap (float): Number from 0 (no overlap) to 1, the fraction of each
+            plot's area covered by the next plot.
+        xlabel (str): Label for the shared x axis.
+        ylabel (str): Label for the y axis (drawn on the bottom plot).
+        ylim (tuple): Lower & upper limits for the y axis of each line.
+        yline (float): Y value at which to draw a horizontal marker for each line.
+        stack_label (str): Label for the stacking axis (shown on the right)
+        stack_ticklabels (array_like): Labels for each line (shown on the right
+            next to the zero line of each plot).
+    """
+    import matplotlib.pyplot as plt
+    import matplotlib.gridspec as grid_spec
+    from matplotlib.transforms import blended_transform_factory
+
+    if isinstance(data, (list, tuple)):
+        data = np.asarray(data)
+    if data.ndim != 2:
+        raise TypeError(f"Expected a 2D array (got {data.ndim}D)")
+
+    if fig is None:
+        fig = plt.figure(figsize=(8, 6), layout="constrained")
+
+    gs = grid_spec.GridSpec(len(data), 1, hspace=-overlap)
+
+    if _isinstance_no_import(data, "xarray", "DataArray"):
+        x_data = data.coords[data.dims[1]]
+        xlabel = xlabel or data.dims[1]
+        stack_label = stack_label or data.dims[0]
+        if stack_ticklabels is None and data.dims[0] in data.coords:
+            stack_ticklabels = data.coords[data.dims[0]].values
+    else:  # Numpy array
+        x_data = np.arange(data.shape[1])
+
+    x_range = np.nanmin(x_data), np.nanmax(x_data)
+    if ylim is not None:
+        y_min, y_max = ylim
+    else:
+        y_min, y_max = np.nanmin(data), np.nanmax(data)
+        if y_min > 0 and (y_max / y_min) > 20:
+            y_min = 0  # Data from just above 0
+        elif y_max < 0 and (y_min / y_max) > 20:
+            y_max = 0  # Data from just below 0
+
+    if yline is None:
+        if y_min <= 0 <= y_max:
+            yline = 0
+        else:
+            yline = np.median(data)
+
+    for i, trace in enumerate(data):
+        ax = fig.add_subplot(gs[i:i + 1, 0:])
+        ax.patch.set_alpha(0)  # Transparent background
+        ax.spines["top"].set_visible(False)
+        ax.spines["right"].set_visible(False)
+
+        ax.plot(x_data, trace)
+
+        # Draw a light line to mark each separate dataset
+        ax.axhline(yline, color='0.7', linewidth=1., zorder=0)
+
+        # Use the same scale on each axes
+        ax.set_ylim(y_min, y_max)
+        ax.set_xlim(*x_range)
+
+        if i < len(data) - 1:
+            # Clear stuff from all but the bottom axes
+            ax.set_xticklabels([])
+            ax.set_xticks([])
+            ax.set_yticks([])
+            ax.set_yticklabels([])
+            ax.spines['left'].set_visible(False)
+            ax.spines['bottom'].set_visible(False)
+        else:
+            # Bottom axes
+            ax.set_xlabel(xlabel)
+            ax.set_ylabel(ylabel)
+
+        # the x coords of this transformation are axes, and the y coords are data
+        if stack_ticklabels is not None:
+            trans = blended_transform_factory(ax.transAxes, ax.transData)
+            ax.text(1.02, yline, str(stack_ticklabels[i]), ha="left", va="center", transform=trans)
+
+    if stack_label:
+        fig.supylabel(stack_label, x=1., ha="right")
+    return fig